etl/polarsetl/requirements.txt
adlfs==2024.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
azure-core==1.35.0
azure-datalake-store==0.0.53
azure-identity==1.23.1
azure-storage-blob==12.26.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
deltalake==0.25.5
frozenlist==1.7.0
fsspec==2025.7.0
idna==3.10
isodate==0.7.2
lxml==5.4.0
msal==1.33.0
msal-extensions==1.3.1
multidict==6.6.3
polars==1.30.0
propcache==0.3.2
pyarrow==20.0.0
pycparser==2.22
PyJWT==2.10.1
requests==2.32.4
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
